feat(shared,analytics-engine): move some shared stuff from analytics engine to shared to avoid workspace dep in shared
