+++ packages/common/tsconfig.json
@@ -14,17 +14,16 @@
     ]
   },
   "include": [
-    "**/*"
-, "../shared/processors/drive-analytics/index.ts", "../shared/processors/drive-analytics/types.ts"  ],
+    "**/*",
+    "../shared/processors/drive-analytics/index.ts",
+    "../shared/processors/drive-analytics/types.ts"
+  ],
   "exclude": [
     "dist",
     "node_modules",
     "storybook-static"
   ],
   "references": [
-    {
-      "path": "../analytics-engine/core"
-    },
     {
       "path": "../config"
     },
