+++ packages/shared/tsconfig.json
@@ -8,12 +8,5 @@
   // include all files in this package's root
   "include": [
     "**/*"
-  ],
-  // add monorepo package paths here
-  // whenever you add "workspace:*" to this package's package.json, also add the path here
-  "references": [
-    {
-      "path": "../analytics-engine/core"
-    }
   ]
 }
